- [wip] acceptance spec - add created file events on file watcher startup (and spec for that) - add startup info

+++ src/main/java/io/github/pgagala/GitService.java
@@ -95,6 +95,9 @@ class GitService {
     //TODO throw or stay with response ?
     Response deleteRepository() {
         try {
+            if(!gitRepositoryLocalFile.exists()) {
+                return Response.success();
+            }
             FileUtils.forceDelete(gitRepositoryLocalFile);
             return Response.success();
         } catch (IOException exc) {
